+++ viadot/tasks/outlook.py
@@ -23,7 +23,6 @@ class OutlookToDF(Task):
         *args: List[Any],
         **kwargs: Dict[str, Any],
     ):
-
         self.mailbox_name = mailbox_name
         self.start_date = start_date
         self.end_date = end_date
@@ -48,6 +47,7 @@ class OutlookToDF(Task):
         mailbox_name: str,
         start_date: str = None,
         end_date: str = None,
+        outbox_list: List[str] = ["Sent Items"],
         limit: int = 10000,
     ) -> pd.DataFrame:
         """
@@ -57,6 +57,8 @@ class OutlookToDF(Task):
             mailbox_name (str): Mailbox name.
             start_date (str, optional): A filtering start date parameter e.g. "2022-01-01". Defaults to None.
             end_date (str, optional): A filtering end date parameter e.g. "2022-01-02". Defaults to None.
+            outbox_list (List[str], optional): List of outbox folders to differenciate between
+                Inboxes and Outboxes. Defaults to ["Sent Items"].
             limit (str, optional): A limit to access last top messages. Defaults to 10_000.
 
         Returns:
@@ -69,9 +71,11 @@ class OutlookToDF(Task):
             end_date=end_date,
             limit=limit,
         )
-        df = outlook.to_df()
+        df = outlook.get_all_mails_to_df(outbox_list=outbox_list)
+        outlook.to_csv(df)
 
         logger.info(
             f"Downloaded the data from the '{outlook.mailbox_name}' into the Data Frame."
         )
+
         return df