+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  `GetSalesOrderData` endpoints in `BusinessCore()` source.
 - Added `url` parameter to `CustomerGauge` source, and `endpoint_url` parameter to `CustomerGaugeToDF` task 
 and `CustomerGaugeToADLS` flow. This parameter enables to pass the endpoint URL by user.
+- Added new parameter `outbox_list` at all leves in `Outlook` connector to tag mailbox folders.
+
+### Changed
+- Changed loop when retrieving email in `Outlook` source file, to cover all possible folders and subfolders.
 
 
 ## [0.4.15] - 2023-05-11